Improvement/script events listiners (#90)
* client creation after client deletion

* client created on client deletion

* Seperate hooks to register events

* removed the isHelloUser check for now

---------
