+++ config/helloApi.ts
@@ -81,7 +81,7 @@ export async function getAllChannels(): Promise<any> {
       {
         headers: {
           authorization: (unique_id || mail || number)
-            ? getLocalStorage('WidgetId')
+            ? getLocalStorage('WidgetId') && getLocalStorage('k_clientId') ? `${getLocalStorage('WidgetId')}:${getLocalStorage('k_clientId')}` : getLocalStorage('WidgetId')
             : (getLocalStorage('a_clientId')
               ? `${getLocalStorage('WidgetId')}:${getLocalStorage('a_clientId')}`
               : getLocalStorage('WidgetId')),
@@ -91,8 +91,8 @@ export async function getAllChannels(): Promise<any> {
 
     if (unique_id || mail || number) {
       setLocalStorage('k_clientId', response?.data?.uuid)
-    } else if (response?.data?.customer_name) {
-      setLocalStorage('default_client_created', 'true')
+    } else {
+      setLocalStorage('a_clientId', response?.data?.uuid)
     }
 
     // Update userData with customer details from response if available
