+++ utils/helloInterceptor.ts
@@ -12,32 +12,32 @@ axios.interceptors.response.use(
   (response) => {
     return response;
   },
-  async function (error) {
-    if (error?.response?.data?.message?.is_exists === 0) {
-      let is_anon = ''
-      if (getLocalStorage('k_clientId')) {
-        is_anon = 'false'
-      } else {
-        is_anon = 'true'
-      }
+  // async function (error) {
+  //   if (error?.response?.data?.message?.is_exists === 0) {
+  //     let is_anon = ''
+  //     if (getLocalStorage('k_clientId')) {
+  //       is_anon = 'false'
+  //     } else {
+  //       is_anon = 'true'
+  //     }
 
-      store.dispatch(setDataInAppInfoReducer({
-        subThreadId: '',
-        currentChannelId: '',
-        currentChatId: '',
-        currentTeamId: ''
-      }))
+  //     store.dispatch(setDataInAppInfoReducer({
+  //       subThreadId: '',
+  //       currentChannelId: '',
+  //       currentChatId: '',
+  //       currentTeamId: ''
+  //     }))
 
-      setLocalStorage('is_anon', is_anon)
-      setLocalStorage('a_clientId', '')
-      setLocalStorage('k_clientId', '')
-      setLocalStorage('default_client_created', 'false')
-      const currentWidgetId = getLocalStorage('WidgetId') || ''
-      setLocalStorage('WidgetId', currentWidgetId)
+  //     setLocalStorage('is_anon', is_anon)
+  //     setLocalStorage('a_clientId', '')
+  //     setLocalStorage('k_clientId', '')
+  //     setLocalStorage('default_client_created', 'false')
+  //     const currentWidgetId = getLocalStorage('WidgetId') || ''
+  //     setLocalStorage('WidgetId', currentWidgetId)
 
-    }
-    return Promise.reject(error);
-  }
+  //   }
+  //   return Promise.reject(error);
+  // }
 );
 
 export default axios;
